rename utils + update frontend
